View snapshot when view is 1024X702
This test app pushes a view controller onto a
navigationviewcontroller’s stack. The pushed viewController has a sub
view which uses core graphics to render a rectangle inset by 10 points.
When the user presses the Camera button, a snapshot is taken of this
subview and placed in the Camera Roll